+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 ### Updated
 1. Fixed initial round of _staticcheck_ lint errors and added _staticcheck_ to
    CI build.
+2. Added file lock to `get-acl` and `compare-acl` commands with `--lockfile` command line option
+   to optionally set the lockfile filepath.
 
 
 ## [0.8.3](https://github.com/uhppoted/uhppoted-app-wild-apricot/releases/tag/v0.8.3) - 2022-12-16